# [10.0][IMP] contract: Performance boost 🚀 