@rafaelbn 2019-01-24
@sbejaoui maybe you are curious about this as you are working for v12 contracts