@sbejaoui 2019-01-24
Thank @rafaelbn  for the ping, it's a good improvement indeed.

@ThomasBinsfeld , @sbidoul